requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
exceptiongroup==1.1.3
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
numpy==1.25.2
oauthlib==3.2.2
outcome==1.2.0
pandas==2.1.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygsheets==2.0.6
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.10.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wsproto==1.2.0
